Move internal/ to pkg/
