+++ pkg/sat/constraints.go
