+++ pkg/sat/constraints_test.go
