+++ pkg/sat/lit_mapping.go
