+++ pkg/sat/search_test.go
