+++ pkg/sat/solve_test.go
