+++ pkg/sat/variable.go
